+++ cypress/integration/EnrollmentPage/StagesAndEventsWidget/index.js
@@ -170,3 +170,27 @@ Then('the sorted list by Report date asc should be displayed', () => {
                 .should('exist');
         });
 });
+
+When(/^you click button Go to full (.*)$/, (stageName) => {
+    cy.get('[data-test="stages-and-events-widget"]').within(() => {
+        cy.get('[data-test="view-all-button"]').contains(stageName).should('exist');
+        cy.get('[data-test="view-all-button"]').contains(stageName).click();
+        cy.wait(100);
+    });
+});
+
+Then('you should navigate to Program Stage list page', () => {
+    cy.url().should('eq', `${Cypress.config().baseUrl}/#/enrollment/stageEvents?programId=WSGAb5XwJ3Y&orgUnitId=DwpbWkiqjMy&stageId=edqlbukwRfQ`);
+});
+
+When(/^you click New (.*)$/, (stageName) => {
+    cy.get('[data-test="stages-and-events-widget"]').within(() => {
+        cy.get('[data-test="create-new-button"]').contains(stageName).should('exist');
+        cy.get('[data-test="create-new-button"]').contains(stageName).click();
+        cy.wait(100);
+    });
+});
+
+Then('you should navigate to Add new page', () => {
+    cy.url().should('eq', `${Cypress.config().baseUrl}/#/enrollmentEventNew?programId=WSGAb5XwJ3Y&orgUnitId=DwpbWkiqjMy&teiId=yFcOhsM1Yoa&enrollmentId=ek4WWAgXX5i&stageId=edqlbukwRfQ`);
+});