+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/StageDetail/hooks/helpers.js
@@ -4,6 +4,7 @@ import type { ApiTEIEvent } from 'capture-core/events/getEnrollmentEvents';
 import { statusTypes, translatedStatusTypes } from '../../../../../../metaData';
 import { getSubValues } from '../../getEventDataWithSubValue';
 import type { StageDataElement } from '../../../../types/common.types';
+import { Comments } from '../Comments.component';
 
 const isEventOverdue = (event: ApiTEIEvent) => moment(event.dueDate).isSameOrBefore(new Date())
     && event.status === statusTypes.SCHEDULE;
@@ -39,6 +40,8 @@ const convertStatusForView = (event: ApiTEIEvent) => {
     };
 };
 
+const convertCommentForView = (event: ApiTEIEvent) => <Comments event={event} />;
+
 const groupRecordsByType = async (events: Array<ApiTEIEvent>, dataElements: Array<StageDataElement>) => {
     // $FlowFixMe
     const dataElementsByType = events.reduce((acc, event) => {
@@ -67,6 +70,7 @@ export {
     isEventOverdue,
     getEventStatus,
     convertStatusForView,
+    convertCommentForView,
     getValueByKeyFromEvent,
     groupRecordsByType,
 };