+++ src/core_modules/capture-core/components/WorkingLists/EventWorkingLists/ReduxProvider/eventWorkingListsReduxProvider.types.js
@@ -24,7 +24,7 @@ import type {
     UnloadingContext,
     UpdateFilter,
     WorkingListTemplate,
-} from '../../WorkingLists';
+} from '../../WorkingListsBase';
 import type {
     AddTemplate,
     DeleteTemplate,