+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsBase/ContextBuilder/ContextProviders/workingListsListViewLoaderContextProvider.types.js
@@ -1,0 +1,30 @@
+// @flow
+import type {
+    ColumnConfigs,
+    LoadView,
+    UpdateList,
+    CancelLoadView,
+    Categories,
+    LoadedViewContext,
+} from '../../workingListsBase.types';
+import type {
+    FiltersData,
+} from '../../../../ListView';
+
+export type Props = $ReadOnly<{|
+    sortById?: string,
+    sortByDirection?: string,
+    filters?: FiltersData,
+    columns: ColumnConfigs,
+    loading: boolean,
+    onLoadView: LoadView,
+    loadViewError?: string,
+    onUpdateList: UpdateList,
+    onCancelLoadView?: CancelLoadView,
+    orgUnitId: string,
+    categories?: Categories,
+    dirtyView: boolean,
+    loadedViewContext: LoadedViewContext,
+    viewPreloaded?: boolean,
+    children: React$Node,
+|}>;